refactor(redshift): rename utils to messageUtils
