+++ src/redshift/messageUtils.ts
@@ -43,7 +43,7 @@ export function createLogsConnectionMessage(warehouseIdentifier: string, error: 
  * @param error
  */
 export function showViewLogsFetchMessage(fetchType: string, identifier: string, error: Error) {
-    const message = `Redshift: Failed to fetch ${fetchType} for ${identifier}: ${(error as Error).message}`
+    const message = `Redshift: failed to fetch ${fetchType} for ${identifier}: ${(error as Error).message}`
     getLogger().error(message)
     showViewLogsMessage(message)
 }
